+++ tests/TransfersTests/TestNodeRpcProxy.cpp
@@ -1,6 +1,11 @@
-// Copyright (c) 2017-2018, The Alloy Developers.
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+/*
+ * Copyright (c) 2017-2018, The Alloy Developers.
+ *
+ * This file is part of Alloy.
+ *
+ * This file is subject to the terms and conditions defined in the
+ * file 'LICENSE', which is part of this source code package.
+ */
 
 #include "gtest/gtest.h"
 
@@ -74,8 +79,8 @@ namespace {
     ASSERT_TRUE(mineBlocks(*nodeDaemons[NODE_0], wallet1.address(), 1));
     ASSERT_TRUE(mineBlocks(*nodeDaemons[NODE_0], wallet1.address(), m_currency.minedMoneyUnlockWindow()));
 
-    wallet1.waitForSynchronizationToHeight(static_cast<uint32_t>(m_currency.minedMoneyUnlockWindow()) + 1);
-    wallet2.waitForSynchronizationToHeight(static_cast<uint32_t>(m_currency.minedMoneyUnlockWindow()) + 1);
+    wallet1.waitForSynchronizationToHeight(m_currency.minedMoneyUnlockWindow() + 1);
+    wallet2.waitForSynchronizationToHeight(m_currency.minedMoneyUnlockWindow() + 1);
 
     PoolChangedObserver observer;
     node0->addObserver(&observer);
